backend/requirements.txt
Django>=5.1.7
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
dj-database-url>=2.1.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
whitenoise>=6.6.0

# Development tools
black>=24.2.0
flake8>=7.0.0
mypy>=1.8.0
django-stubs>=4.2.7
types-requests>=2.31.0.20240218 